+++ spring-cloud-starter-tencent-polaris-router/src/test/java/com/tencent/cloud/polaris/router/RouterUtilsTest.java
@@ -74,6 +74,8 @@ public class RouterUtilsTest {
 					.thenReturn(testNamespaceAndService);
 			MetadataContext metadataContext = Mockito.mock(MetadataContext.class);
 			mockedMetadataContextHolder.when(MetadataContextHolder::get).thenReturn(metadataContext);
+			Mockito.when(metadataContext.getContext(anyString(), anyString(), anyString()))
+					.thenReturn(testNamespaceAndService);
 
 			int instanceSize = 100;
 			int weight = 50;