+++ CHANGELOG.md
@@ -48,3 +48,4 @@
 - [feat: implement circuit breaker in enhance plugin, support listen config group, support refresh single config in refresh_context mode.](https://github.com/Tencent/spring-cloud-tencent/pull/1501)
 - [feat:support polaris event.](https://github.com/Tencent/spring-cloud-tencent/pull/1502)
 - [feat:support circuit breaker metrics reporting.](https://github.com/Tencent/spring-cloud-tencent/pull/1503)
+- [feat: support gateway context, feign eager-load support default value.](https://github.com/Tencent/spring-cloud-tencent/pull/1504)